+++ lib/td_core/search/index_worker.ex
@@ -1,112 +1,33 @@
 defmodule TdCore.Search.IndexWorker do
   @moduledoc """
-  GenServer to run reindex task
+  Behaviour for index tasks
   """
 
-  @behaviour TdCache.EventStream.Consumer
+  @type index :: atom()
+  @type ids :: list(number())
 
-  use GenServer
+  @callback start_link(index) :: {:ok, any()} | {:error, any()}
+  @callback reindex(index, ids) :: :ok
+  @callback delete(index, ids) :: term()
+  @callback get_index_workers() :: list()
 
-  alias TdCore.Search.Indexer
-  alias TdCore.Search.MockIndexWorker
-  alias TdCore.Utils.Timer
+  def start_link(index),
+    do: impl().start_link(index)
 
-  require Logger
+  def reindex(index, ids),
+    do: impl().reindex(index, ids)
 
-  ## Public API that maybe uses test
+  def delete(index, ids),
+    do: impl().delete(index, ids)
 
-  def start_link(index) do
-    case Application.get_env(:td_core, :env) do
-      :test -> MockIndexWorker.start_link(index)
-      _ -> GenServer.start_link(__MODULE__, index, name: index)
+  def get_index_workers,
+    do: impl().get_index_workers()
+
+  defp impl() do
+    if Application.get_env(:td_core, :env) == :test do
+      TdCore.Search.IndexWorkerMock
+    else
+      TdCore.Search.IndexWorkerImpl
     end
-  end
-
-  def reindex(index, ids) do
-    case Application.get_env(:td_core, :env) do
-      :test -> MockIndexWorker.reindex(index, ids)
-      _ -> GenServer.cast(index, {:reindex, ids})
-    end
-  end
-
-  def delete(index, ids) do
-    case Application.get_env(:td_core, :env) do
-      :test -> MockIndexWorker.delete(index, ids)
-      _ -> GenServer.call(index, {:delete, ids})
-    end
-  end
-
-  def get_index_workers do
-    :td_core
-    |> Application.get_env(TdCore.Search.Cluster, [])
-    |> Keyword.fetch!(:aliases)
-    |> Map.keys()
-    |> Enum.map(&Supervisor.child_spec({TdCore.Search.IndexWorker, &1}, id: &1))
-  end
-
-  ## EventStream.Consumer Callbacks
-
-  @impl TdCache.EventStream.Consumer
-  def consume(events) do
-    index_scope = get_index_template_scope()
-
-    Enum.map(events, fn
-      %{event: "template_updated", scope: scope} ->
-        Map.get(index_scope, String.to_atom(scope))
-
-      _ ->
-        nil
-    end)
-    |> Enum.reject(&is_nil(&1))
-    |> Enum.uniq()
-    |> Enum.each(&reindex(&1, :all))
-
-    :ok
-  end
-
-  ## GenServer Callbacks
-
-  @impl GenServer
-  def init(index) do
-    Logger.info("Running IndexWorker for index #{index}")
-    {:ok, index}
-  end
-
-  @impl GenServer
-  def handle_cast({:reindex, ids}, index) do
-    Logger.info("Started indexing for #{index}")
-
-    Timer.time(
-      fn -> Indexer.reindex(index, ids) end,
-      fn millis, _ -> Logger.info("#{index} indexed in #{millis}ms") end
-    )
-
-    {:noreply, index}
-  end
-
-  @impl GenServer
-  def handle_call({:delete, ids}, _from, index) do
-    reply =
-      Timer.time(
-        fn -> Indexer.delete(index, ids) end,
-        fn millis, _ -> Logger.info("Rules deleted in #{millis}ms") end
-      )
-
-    {:reply, reply, index}
-  end
-
-  defp get_indexes(module) do
-    module
-    |> Application.get_env(TdCore.Search.Cluster, [])
-    |> Keyword.fetch!(:indexes)
-  end
-
-  defp get_index_template_scope do
-    :td_core
-    |> get_indexes()
-    |> Enum.map(fn {index, resource} ->
-      {Map.get(resource, :template_scope), index}
-    end)
-    |> Map.new()
   end
 end