+++ test/td_core/search/index_worker_test.ex
@@ -1,0 +1,8 @@
+defmodule TdCore.Search.IndexWorkerTest do
+  use ExUnit.Case
+
+  import ExUnit.CaptureLog
+  require Logger
+
+  alias TdCore.Search.IndexWorker
+end